Move domain proxy to routes
This matches better with the other routes.

Also add a missing authentication check to the path proxy web socket.
