+++ src/node/routes/index.ts
@@ -13,12 +13,12 @@ import { rootPath } from "../constants"
 import { Heart } from "../heart"
 import { replaceTemplates } from "../http"
 import { loadPlugins } from "../plugin"
-import * as domainProxy from "../proxy"
 import { getMediaMime, paths } from "../util"
 import { WebsocketRequest } from "../wsRouter"
+import * as domainProxy from "./domainProxy"
 import * as health from "./health"
 import * as login from "./login"
-import * as proxy from "./proxy"
+import * as proxy from "./pathProxy"
 // static is a reserved keyword.
 import * as _static from "./static"
 import * as update from "./update"
